+++ src/index.js
@@ -2,6 +2,7 @@ import ValueStream from './ValueStream';
 import ValueStreamFast from './ValueStreamFast';
 import ValueMapStream from './ValueMapStream';
 import ValueMapStreamFast from './ValueMapStreamFast';
+import ValueObjectStream from './ValueObjectStream';
 import Event, { EventFilter } from './Event';
 import setProxy from './setProxy';
 import * as constants from './constants';
@@ -17,6 +18,7 @@ export default {
   ValueStreamFast,
   ValueMapStream,
   ValueMapStreamFast,
+  ValueObjectStream,
   setProxy,
   addTrans,
 };